Diego

+++ SMedia/SMedia/Models/Community.cs
@@ -7,13 +7,13 @@ namespace SMedia.Models
 {
     public class Community
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Color { get; set; }
         public string Description { get; set; }
         public string? Picture { get; set; }
         public DateTime CreationDate { get; set; }
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public bool Active { get; set; }
 
         public virtual User Owner { get; set; }
