+++ SMedia/SMedia/Models/FollowedUser.cs
@@ -7,9 +7,9 @@ namespace SMedia.Models
 {
     public class FollowedUser
     {
-        public int Id { get; set; }
-        public int FollowerId { get; set; }
-        public int FollowedId { get; set; }
+        public long Id { get; set; }
+        public long FollowerId { get; set; }
+        public long FollowedId { get; set; }
         public DateTime DateOfFollow { get; set; }
 
         public virtual User Follower { get; set; }
